Validando se evento está associado à uma venda na deleção.

+++ src/formularios/DeletarEvento.java
@@ -1,6 +1,7 @@
 package formularios;
 
 import bancodedados.dao.EventoDAO;
+import bancodedados.dao.VendaDAO;
 import java.util.Scanner;
 import negocio.Evento;
 import uteis.Funcoes;
@@ -18,21 +19,26 @@ public class DeletarEvento {
         try {
             Evento evento = new bancodedados.dao.EventoDAO().getEvento(scanner.nextInt()); 
             if(evento.getId() > 0){
-                System.out.println("\n:: ID: " + evento.getId());
-                System.out.println(":: NOME: " + evento.getNome());            
-                System.out.println(":: DATA: " + evento.getDataInicio());
-                System.out.println(":: HORA: " + evento.getHoraInicio());
-                System.out.println(":: LOCAL: " + evento.getEndereco() + "\n");
-                
-                
-                Funcoes.subCabecalho("Confirmação");                
-                System.out.print("Deseja realmente deletar esse evento? [s/n]: ");
-                if(Funcoes.opcSimNao()){
-                    new EventoDAO().removerEvento(evento);
-                    System.out.println("Evento foi removido com exito.");
+                if(!new VendaDAO().isEventoAssociado(evento.getId())){
+                    System.out.println("\n:: ID: " + evento.getId());
+                    System.out.println(":: NOME: " + evento.getNome());            
+                    System.out.println(":: DATA: " + evento.getDataInicio());
+                    System.out.println(":: HORA: " + evento.getHoraInicio());
+                    System.out.println(":: LOCAL: " + evento.getEndereco() + "\n");
+
+
+                    Funcoes.subCabecalho("Confirmação");                
+                    System.out.print("Deseja realmente deletar esse evento? [s/n]: ");
+                    if(Funcoes.opcSimNao()){
+                        new EventoDAO().removerEvento(evento);
+                        System.out.println("Evento foi removido com exito.");
+                    }
+                    else{
+                        System.out.println("Evento foi mantido cadastrado com ID: " + evento.getId());
+                    }
                 }
                 else{
-                    System.out.println("Evento foi mantido cadastrado com ID: " + evento.getId());
+                    System.out.println("Evento associado à uma venda.");
                 }
             }
             else{
